+++ deployment-test/README.md
@@ -1,9 +1,9 @@
 ## Deployment Test
 
 ### Description
-Automated API checks for Rosetta deployments on testnet, stagning, shelley-qa and mary-qa.
+Automated API checks for Rosetta deployments on testnet, stagning, shelley-qa and mary-qa. Can also be used for local instance.
 
 ### Steps:
-- `npm install newman`
-- `nodejs data-deploy-test`
+- `npm i newman`
+- `node data-deployment-test-all-envs`
