+++ deployment-test/data-local-docker-test.js
@@ -1,12 +1,12 @@
 const newman = require('newman');
 
-const envFile = require('some_arg');
+const env = process.argv[2];
 const port = 8081;
 
 newman.run({
     collection: require('./Cardano-Rosetta.postman_collection.json'),
-    environment: require(endFile),
-    envVar: [ { "key":"", "value":`localhost:${port}` }  ]
+    environment: require(`./environments/${env}`),
+    envVar: [ { "key":"", "value":`localhost:${port}` }  ],
     reporters: ['cli', 'json']
 }, function (err) {
 	if (err) { throw err; }
